Stripe webhook delete, webhook secret logic

+++ php/seatreg_functions.php
@@ -2364,37 +2364,27 @@ function seatreg_update() {
 	);
 
 	if( $oldOptions->stripe_payments === '0' && $_POST['stripe-payments'] === 1) {
-		if( !StripeWebhooksService::isStripeWebhookCreated($_POST['stripe-api-key']) ) {
-			//Create a Stripe webhook when turning on Stripe payments
+		//Stripe payments is turned on
+		if( !StripeWebhooksService::isStripeWebhookCreatedForCurrentSite($_POST['stripe-api-key']) ) {
+			//Create a new Stripe webhook
 			$webhook = StripeWebhooksService::createStripeWebhook($_POST['stripe-api-key']);
-
-			$wpdb->update(
-				"$seatreg_db_table_names->table_seatreg_options",
-				array(
-					'stripe_webhook_secret' => $webhook->secret,
-				),
-				array(
-					'registration_code' => sanitize_text_field($_POST['registration_code'])
-				),
-				'%s',
-				'%s'
+			SeatregOptionsService::updateStripeWebhookSecret($webhook->secret, sanitize_text_field($_POST['registration_code']));
+		}else {
+			//webhook already created. Set stripe_webhook secret fro existing webhook
+			SeatregOptionsService::updateStripeWebhookSecret(
+				SeatregOptionsRepository::getActiveStripeWebhookSecret($_POST['stripe-api-key']),
+				sanitize_text_field($_POST['registration_code']),
 			);
-		
 		}
 	}else if( $oldOptions->stripe_payments === '1' &&  $_POST['stripe-payments'] === 0) {
-		//Remove Stripe webhook when Stripe payments are turned off
-		StripeWebhooksService::removeStripeWebhook($_POST['stripe-api-key']);
+		//Turning off Stripe payment
+		$activeStripeKeyCount = SeatregOptionsRepository::getActiveStripeKeyUsage($_POST['stripe-api-key']);
+		SeatregOptionsService::updateStripeWebhookSecret(null, sanitize_text_field($_POST['registration_code']));
 
-		$wpdb->update(
-			"$seatreg_db_table_names->table_seatreg_options",
-			array(
-				'stripe_webhook_secret' => null,
-			),
-			array(
-				'registration_code' => sanitize_text_field($_POST['registration_code'])
-			),
-			'%s'
-		);
+		if( $activeStripeKeyCount === 0 ) {
+			//Remove the webhook if not used anymore
+            StripeWebhooksService::removeStripeWebhook($_POST['stripe-api-key']);
+        }
 	}
 
 	return ($status1 !== false && $status2 !== false);
